combine-diff: zero memory used for callback filepairs
In commit 25e5e2bf85 (combine-diff: support format_callback,
2011-08-19), the combined-diff code learned how to make a multi-sourced
diff_filepair to pass to a diff callback. When we create each filepair,
we do not bother to fill in many of the fields, because they would make
no sense (e.g., there can be no rename score or broken_pair flag because
we do not go through the diffcore filters). However, we did not even
bother to zero them, leading to random values. Let's make sure
everything is blank with xcalloc, just as the regular diff code does.

We would potentially want to set the "status" flag to
something non-zero, but it is not clear to what. Possibly a
new DIFF_STATUS_COMBINED would make sense, as this is not
strictly a modification, nor does it fit any other category.

Since it is not yet clear what callers would want, this
patch simply leaves it as "0", the same empty flag that is
seen when diffcore_std is not used at all.

Signed-off-by: Jeff King <peff@peff.net>