rebase -i --rebase-merges: offer a smart way to rebase merge commits
Previously, we punted on the question how to carry over amendments to
merge commits. Instead, we always performed new merges from scratch.

Such amendments to merge commits may very well be necessary, though,
e.g.  if one side of the history changed a function signature and the
other side added a caller.

In a discussion inspired by Johannes Sixt's "cherry-pick -m 1" idea,
Phillip Wood came up with the fundamental idea behind this commit.

The premise is that both "rebase" and "merge" try to reconcile diverging
changes. Both "rebase" and "merge" should result in identical trees
(after merge conflicts are resolved, if there were any).

A "merge" typically boils down to a "3-way" merge, with a "merge base"
and two "merge heads" that diverged from said merge base. Think e.g.
last week's master as the merge base, from where a developer branched
off a topic branch (the first merge head), and the current master as
second merge head.

A "rebase" is more complicated than a single "3-way merge": When
rebasing commits, they are "cherry-picked" one by one. That way, the
changes introduced by that commit are reconciled with the changes
introduced by the "rebase" so far: a "3-way merge" between HEAD and the
cherry-picked commit, with the latter's parent commit as merge base
(HEAD would be the rebased parent commit in the common case).

In the context of a cherry-pick, it is important to keep in mind that
the "diverging changes" are not reflected by commit history. For the
purpose of a 3-way merge, they don't have to be.

Now let's look again at the problem of "rebasing merge commits", i.e.
how to "cherry-pick a merge commit"? It is not as simple as a "3-way
merge" (a single one, that is) because we do not have a single merge
base from where sprung exactly two diverging changesets.

Just like when we cherry-pick a regular commit, in the case of a merge
commit we have diverging changes for *every* one of its parents. And
just like before, the original parent commit is the merge base, and the
rebased parent commit is a merge head. The other merge head is the
original merge commit itself.

By starting with the original merge commit and then performing these
3-way merges sequentially, one for every parent, we reconcile all
diverging changes. So we "rebased" the merge commit, including all of
its amendments.

This is the most logical generalization of the cherry-pick concept: if
we only have one parent, the strategy outlined above is identical to
Git's cherry-pick operation.

Side-note: While this description only talked about merge commits with
exactly two parent commits above, the principle still holds for merge
commits with more than two parent commits ("octopus merges").

That strategy requires an original merge commit, with a matching number
of parents, though, and therefore it would not make sense in general to
do this for every `merge` command in the todo list: what if there was no
original merge commit, or if the specified merge commit had a different
number of parents? It therefore cannot be the default mode for the
`merge` command, thus we introduce a new flag `-R` for that.

However, for existing merge commits, this strategy *is* valid and *does*
lead to the most intuitive results. Therefore, let's use it by default
when generating todo lists in `git rebase --rebase-merges`.

Original-idea-by: Phillip Wood <phillip.wood@dunelm.org.uk>
Helped-by: Igor Djordjevic <igor.d.djordjevic@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>