commit-graph: not compatible with replace objects
Create new method commit_graph_compatible(r) to check if a given
repository r is compatible with the commit-graph feature. Fill the
method with a check to see if replace-objects exist. Also, check if the
replace objects are enabled with 'read_replace_objects'.

Test this interaction succeeds, including ignoring an existing
commit-graph and failing to write a new commit-graph.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>