Accommodate for forgotten re-commits
In the Git maintainer's workflow, it happens quite frequently that
patches are applied and pushed out multiple times, until they stabilize.
In other words, the same Message-ID may well correspond to one commit in
git.git today, and to another tomorrow.

To handle that, GitGitGadget's `update-mail-to-commit-notes.sh` script
tries to remove the mail-to-commit mapping to stale commits, and to
replace them with up to date mappings.

Unfortunately, an early version of that script contained two bugs, and it
was in operation until the parent commit of the one whose commit message
you, dear reader, are reading right now. The bug was that while it
looks at the diff of the corresponding revisions of
`refs/notes/commit-to-mail`, it did not turn off rename detection (but
it has to, to remove and then add the mappings), and it did not ensure
that removals are handled before adds.

This has been fixed now, and the tree of this here revision corresponds
to the output of a re-run with the fixed script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>