test functions: add function `test_file_not_empty`
test-lib-functions: add a helper function that checks for a file and that
the file is not empey. The helper function will provide better error message
in case of failuer and improve readability

The function `test_file_not_empty`, first calls `test_path_is_file` to
check verify if a file is provided, if it is not then an error message
is printed, short-circuiting the remaining code, if <path> is indeed a
file then check `test -s` is applied to check is size of file is more
than zero, failing which another error message is returned.

Signed-off-by: Rohit Ashiwal <rohit.ashiwal265@gmail.com>