git-completion: sanitize the command names
Do not declare/execute commands that contain invalid or special characters.

If the autocompleted command contains illegal characters, for example when
misspelling `git pull` as `git [ull`, then the user will see an error. This
patch adds a character whitelist for commands that strips all but
lowercase alphabetic characters and dashes, so that misspells fail silently.

This patch uses the `[[` keyword that is not sh-compatible, but it's
okay since the change affects BASH and ZSH-specific autocomplete
scripts.

Signed-off-by: Vladyslav Burzakovskyy <vladislav.burzakovskij@satoshilabs.com>