show --continue/skip etc. consistently in synopsis
The synopsis for am shows the command mode options as
    (--continue | --skip | --abort | --quit)
They are on a single line and in parenthesis as they are not
optional. Fix the merge and rebase docs to match this style.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>
---
This patch is based on top of nd/merge-quit.
The docs for cherry-pick/revert are updated to this style by
https://public-inbox.org/git/20190616082040.9440-1-rohit.ashiwal265@gmail.com/T/#u