psuh: add a built-in by popular demand
Internal metrics indicate this is a command many users expect to be
present. So here's an implementation to help drive customer
satisfaction and engagement: a pony which doubtfully greets the user,
or, a Pony Saying "Um, Hello!!" (PSUH). This is also My First
Contribution.

Signed-off-by: george espinoza <gespinoz2019@gmail.com>

psuh: add a built-in by popular demand

Internal metrics indicate this is a command many users expect to be
present. So here's an implementation to help drive customer
satisfaction and engagement: a pony which doubtfully greets the user,
or, a Pony Saying "Um, Hello" (PSUH). This is also My First
Contribution.

Signed-off-by: george espinoza <gespinoz2019@gmail.com>

psuh: add a built-in by popular demand

Internal metrics indicate this is a command many users expect to be
present. So here's an implementation to help drive customer
satisfaction and engagement: a pony which doubtfully greets the user,
or, a Pony Saying "Um, Hello" (PSUH).
My first commit did not have accurate text from tutorial.

Signed-off-by: george espinoza <gespinoz2019@gmail.com>