[RFC] rebase: fix advice when a fixup creates an empty commit
Add a specific message to advise the user what to do when a fixup or
squash command would create an empty commit.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>
---
I'm slightly nervous of moving the call to determine_whence() to
finalized_deferred_config() I think it is ok but need to do a more
through audit of the code to be sure.