rebase -i: fix --committer-date-is-author-date
Johanathan Nieder reported that `rebase --committer-date-is-author-date`
failed with the error "invalid date format: @@2592000 +0000" when the
backend was switched to use the sequencer [1]. This is because when we
read the date from the author script file is already prefixed by '@' so
we do not need to add one. This did not show up in our tests
because once match_object_header_date() fails to match because of the
extra '@' parse_date_basic() strips the prefix and a date with more than
8 digits is parsed as a number of seconds by match_digit().

While fixing this I also noticed that we were not setting
GIT_COMMITTER_DATE in the environment when forking `git merge`. This was
untested. The tests we did have used commits where the author date was
the same as the current value $GIT_COMMITTER_DATE in the environment so
they did not test that we were actually using the author date when
creating the commit.

As we have already read GIT_AUTHOR_DATE into cmd.env_array we now use
that rather than re-reading the author-script file. I've moved the code
that handles opts->ignore date so that all the code setting the child
environment is together and changed it so we no longer set
GIT_COMMITTER_DATE twice when --ignore-date is combined with
--committer-date-is-author-date.

[1] https://lore.kernel.org/git/<20200110231436.GA24315@google.com>

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>