commit: replace rebase/sequence booleans with single pick_state enum
In 116a408 (commit: give correct advice for empty commit during a rebase,
2019-10-22), the boolean `rebase_in_progress` was introduced by dscho to
handle instances when cherry-pick and rebase were occurring at the same time.
This created a situation where two independent booleans were being used
to define the state of git at a point in time.

Under his recommendation to follow guidance from Junio[1] it was decided
that an `enum` that defines the state of git would be a more effective path
forward.

Removed multiple booleans `rebase_in_progress` and `sequencer_in_use` and
replaced with a single `pick_state` enum that determines if, when
cherry-picking, we are in a rebase, multi-pick, or single-pick state.

This differs from the previous patch version by broadening the options
in `pick_state`, converting double `if` statement to a conditional
operator, and converting the `if/multi-else` statements to a `switch` to
simplify the codebase.

[1] https://lore.kernel.org/git/xmqqr234i2q0.fsf@gitster-ct.c.googlers.com/

Signed-off-by: Ben Curtis <nospam@nowsci.com>