git: advice user when explicitly setting git-dir without work-tree
git --git-dir <path> is a bit confusing and sometimes doesn't work as
the user would expect it to.

For example, if the user runs `git --git-dir=<path> status`, git
will not be able to figure out the work tree path on its own and
will assign the work tree to the user's current work directory.
When this assignment is wrong, then the output will not match the
user's expectations.

This patch displays a hint message when --git-dir is set
without the --work-tree advising the user to use git -C instead.
It also updates the documentation to make it clearer.

Signed-off-by: Heba Waly <heba.waly@gmail.com>