sequencer.c: define a file path to store hash of a commit
This file stores info of the commit at a pick before a pick-fixup/squash
chain eg. for this series of commands

              p1---p---f1---s---f....

The commit info for HEAD corresponding to pick p1 will be stored in file
here. Although the name suggests bad_fixup it is actually pick p that
has merge conflicts with the base here and we want to avoid applying
fixup to f1 onto p1 if p is skipped.

Currently if pick p results in a conflict, and is skipped f1 gets
fixed up to p1.

Signed-off-by: Parth Gala <parthpgala@gmail.com>