sequencer.c: write commit hash to file rebase_merge/avoid_bad_fixup
The condition here in the 'else if' statement verifies that the commit
whose hash is written to file is actually that of pick p1 since we are
writing before the call to do_pick_commit, refer f7f082c9bf
p could also be an edit or reword but not fixup or squash

            p1---p---f1---f---f...

The writing to file is done based upon the implementation of the
`intend_to_amend` function in sequencer.c which similarly copies a
single hash to a file.

The `if` ensures that we are dealing with a fixup whose preceding
pick/edit/reword had merge conflicts and was skipped. The file contents
are compared with the current head position and if the two are same it
means that the conflicting pick was skipped. Hence we error out without
amending HEAD.

Signed-off-by: Parth Gala <parthpgala@gmail.com>