tests: get rid of the transitional prereq
We introduced the `PREPARE_FOR_MAIN_BRANCH` prereq for the sole purpose
of allowing us to perform the non-trivial adjustments regarding the
`master` -> `main` rename before the automatable ones.

We already removed it from all tests except from t9902 (which took a
little longer to transition to using `main` as default branch name). Now
that the transition is complete, we do not need that prereq anymore.

In addition to the manual removal of the prereq from `t/test-lib.sh`,
this patch includes the result of this command, looking at the _entire_
test suite, just to make sure:

	sed -i 's/PREPARE_FOR_MAIN_BRANCH[ ,]//' t/t[0-9]*.sh

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>