contrib: adjust to the new default branch name
Since we just changed the default branch name to `main`, a couple of
places in contrib/ need to be adjusted, most notably the tests for the
`git subtree` command.

Note: there are quite a few matches for the pattern `master` in
`contrib/`, still; These are mostly URLs for other projects, or
historical references, and are therefore left untouched.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>