diff --color-moved-ws=allow-indentation-change: simplify and optimize
If we already have a block of potentially moved lines then as we move
down the diff we need to check if the next line of each potentially
moved line matches the current line of the diff. The implementation of
--color-moved-ws=allow-indentation-change was needlessly performing
this check on all the lines in the diff that matched the current line
rather than just the current line. To exacerbate the problem finding
all the other lines in the diff that match the current line involves a
fuzzy lookup so we were wasting even more time performing a second
comparison to filter out the non-matching lines. Fixing this reduces
time to run
  git diff --color-moved-ws=allow-indentation-change v2.28.0 v2.29.0
by 93% compared to master and simplifies the code.

Test                                                                 master              HEAD
---------------------------------------------------------------------------------------------------------------
4002.1: diff --no-color-moved --no-color-moved-ws large change        0.51( 0.45+0.06)   0.52(0.45+0.06) +2.0%
4002.2: diff --color-moved --no-color-moved-ws large change           1.14( 1.09+0.04)   1.14(1.10+0.04) +0.0%
4002.3: diff --color-moved-ws=allow-indentation-change large change  19.25(19.17+0.06)   1.25(1.19+0.06) -93.5%
4002.4: log --no-color-moved --no-color-moved-ws                      1.56( 1.42+0.14)   1.56(1.39+0.16) +0.0%
4002.5: log --color-moved --no-color-moved-ws                         1.76( 1.65+0.10)   1.76(1.62+0.13) +0.0%
4002.6: log --color-moved-ws=allow-indentation-change                 2.33( 2.21+0.11)   1.79(1.64+0.15) -23.2%

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>