diff --color-moved-ws=allow-indentation-change: improve hash lookups
As libxdiff does not have a whitespace flag to ignore the indentation
the code for --color-moved-ws=allow-indentation-change uses
XDF_IGNORE_WHITESPACE and then filters out any hash lookups where
there are non-indentation changes. This filtering is inefficient as
we have to perform another string comparison.

By using the offset data that we have already computed to skip the
indentation we can avoid using XDF_IGNORE_WHITESPACE and safely remove
the extra checks which improves the performance by 13% and paves the
way for the elimination of string comparisons in the next commit.

This change slightly increases the runtime of other --color-moved
modes. This could be avoided by using different comparison functions
for the different modes but after the next two commits there is no
measurable benefit. The increased runtime for --no-color-moved
--no-color-moved-ws is strange and I can't explain it, it does seem
to be repeatable on the laptop I ran the benchmarks on though. On a
desktop system with an older processor there is no such increase.

Test                                                                  HEAD^             HEAD
--------------------------------------------------------------------------------------------------------------
4002.1: diff --no-color-moved --no-color-moved-ws large change        0.49(0.44+0.05)   0.51(0.47+0.04) +4.1%
4002.2: diff --color-moved --no-color-moved-ws large change           1.09(1.02+0.06)   1.14(1.09+0.05) +4.6%
4002.3: diff --color-moved-ws=allow-indentation-change large change   1.21(1.14+0.06)   1.05(1.00+0.04) -13.2%
4002.4: log --no-color-moved --no-color-moved-ws                      1.40(1.27+0.12)   1.56(1.43+0.12) +11.4%
4002.5: log --color-moved --no-color-moved-ws                         1.60(1.48+0.11)   1.76(1.62+0.14) +10.0%
4002.6: log --color-moved-ws=allow-indentation-change                 1.64(1.53+0.10)   1.75(1.64+0.10) +6.7%

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>