p4 tests: avoid relying on a specific default branch name
To allow for changing the default branch name used by `git init`, let's
not rely on the initial branch name to be called `master`.

In the git-p4 tests, this needs a little bit of care because the
pseudo-remote branch `p4/master` needs to be left untouched by this
patch: `git p4` will use that remote name, at least for the time being,
and not look at `init.defaultBranch` at all (because why would it?).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>