git-mv: fix git mv bug with case insensitive fs
On a case-insensitive fs, where core.caseInsensitive=true, attempting
to git mv a file already in the index under a new casing will assert.
Since 9b906af657 the check that git mv does to ensure
the src is in the cache respects caseInsensitive. As a result git mv
allows a move from a file that has a different case in the index than
it does on disk. After the rename on disk, git mv fails to find the
file in the cache in order to rename it in the index, and asserts.

This is the simplest possible fix, suggested by @tboegi. It does
leave the file renamed on disk, but that is easy to reverse after
the error.

Signed-off-by: Dan Moseley <danmose@microsoft.com>