ci(vs-build): use actions/cache@v2.1.4 to work around the backslash problem
There is currently a problem with actions/cache@v2: when specifying a
path that contains slashes (such as `compat/vcbuild/vcpkg`, which we now
do in the CI/PR builds), on Windows they are converted to backslashes
_and_ it uses the `tar.exe` found in the PATH.

And since we install git-sdk-64-minimal, there is a GNU tar in the PATH
that thinks that backslashes are a perfectly fine file name characters,
not directory separators (the MSYS2 runtime will translate file name
characters that are legal on Linux/Unix but illegal on Windows, mapping
them into the private Unicode page). As a consequence, it fails to tar
up the vcpkg files.

Version 2.1.4 of actions/cache fixes these problems, by converting the
backslashes back to forward slashes _and_ by insisting on using the
`tar.exe` found in `C:\Windows\system32`.

So: problem solved, right?

Not so fast. An _unrelated_ bug was introduced in that version
(actions/checkout@v2.1.4 now uses GNU tar on macOS, which seems to be
unable to restore certain data shapes that were previously cached using
BSD tar, see https://github.com/actions/cache/issues/527 for details)
and therefore v2 was rolled back to point to v2.1.3 for the moment.

So let's hard-code v2.1.4 for now, hoping that the issue will be
resolved soon.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>