approximate-picks: new API to see if commits are cherry-picks/reverts
We want some functions to check if commits are cherry-picks or reverts,
based on information recorded in the commit message.  Namely, substrings
of the form
  This reverts commit ... [reversing\nchanges made to ...]
  cherry picked from commit ...
When these strings are found, get_message_pick() can return that the
specified commit is a cherry-pick or revert and what other commit the
specified commit is a cherry-pick or revert of.

We will use this later to implement --remerge-diff-only, where normal
non-merge commits that are not cherry-picks or reverts are omitted from
log/show output, and cherry-picks or reverts are shown as a diff against
an automatic re-pick/re-revert.  Thus, the new flag will be a way to see
what changes the creator of the commit made in order to complete the
cherry-pick or revert, such as removing conflict markers.

Signed-off-by: Elijah Newren <newren@gmail.com>