[GSOC] ref-filter: add deref member to struct used_atom
In the original logic, the dereference flag '*' is stored
at the beginning of used_atom[i].name. (e.g. '*objectname:short')
This leads us to make some check on used_atom[i].name[0] to
know if we the atom needs to be dereferenced.

So add `deref` flag to struct used_atom used to indicate
whether the atom needs to be dereferenced and let used_atom
record only the remaining part (i.e. 'objectname:short')

This can simplify the logic of the program.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Hariom Verma <hariom18599@gmail.com>
Signed-off-by: ZheNing Hu <adlternative@gmail.com>