[GSOC] ref-filter: reuse object content
When we are using %(raw) atom, we get the object's content
in get_object(), and store it in oi.content, this is a
dynamically allocated memory. But in grab_sub_body_contents(),
we allocated a new piece of memory to intermediate data `v->s`,
and copy the oi->cotent to it. This has a lot of memory allocation
overhead and copy overhead.

It is worth noting that parse_object_buffer() in get_object()
may "eaten" the buffer, which this means git backend will cache the
buffer and we lost the control of this buffer.

So pass the "eaten" variable from get_object() to
grab_sub_body_contents(). If the buffer have been "eaten",
we copy the memory, otherwise we can reuse this buffer.

When we skip parse_object_buffer(), this buffer is obviously
not been "eaten", so we can reuse it with confidence.

This can give git cat-file --batch a 6% performance
improvement. Now the performance of git cat-file --batch is
2% better than before reusing ref-filter logic.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Hariom Verma <hariom18599@gmail.com>
Signed-off-by: ZheNing Hu <adlternative@gmail.com>